+++ bigo-druid-handler/src/java/org/apache/hadoop/hive/druid/extension/cardinality/accurate/collector/LongRoaringBitmapCollectorFactory.java
@@ -17,11 +17,11 @@
  * under the License.
  */
 
-package org.apache.hadoop.hive.druid.extension.accurate.collector;
+package org.apache.hadoop.hive.druid.extension.cardinality.accurate.collector;
 
 
-import org.apache.hadoop.hive.druid.extension.accurate.bitmap64.LongBitmapFactory;
-import org.apache.hadoop.hive.druid.extension.accurate.bitmap64.LongRoaringBitmapFactory;
+import org.apache.hadoop.hive.druid.extension.cardinality.accurate.bitmap64.LongBitmapFactory;
+import org.apache.hadoop.hive.druid.extension.cardinality.accurate.bitmap64.LongRoaringBitmapFactory;
 
 import java.nio.ByteBuffer;
